+++ pkg/analyzer/test/src/diagnostics/generic_function_type_cannot_be_bound_test.dart
@@ -72,26 +72,29 @@ class GenericFunctionTypeCannotBeBoundWithoutGenericMetadataTest
 
   test_class() async {
     await assertErrorsInCode(r'''
+// @dart=2.12
 class C<T extends S Function<S>(S)> {
 }
 ''', [
-      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 18, 16),
+      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 32, 16),
     ]);
   }
 
   test_genericFunction() async {
     await assertErrorsInCode(r'''
+// @dart=2.12
 late T Function<T extends S Function<S>(S)>(T) fun;
 ''', [
-      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 26, 16),
+      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 40, 16),
     ]);
   }
 
   test_genericFunctionTypedef() async {
     await assertErrorsInCode(r'''
+// @dart=2.12
 typedef foo = T Function<T extends S Function<S>(S)>(T t);
 ''', [
-      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 35, 16),
+      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 49, 16),
     ]);
   }
 
@@ -103,9 +106,10 @@ class C<T extends void Function(S Function<S>(S))> {}
 
   test_typedef() async {
     await assertErrorsInCode(r'''
+// @dart=2.12
 typedef T foo<T extends S Function<S>(S)>(T t);
 ''', [
-      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 24, 16),
+      error(CompileTimeErrorCode.GENERIC_FUNCTION_TYPE_CANNOT_BE_BOUND, 38, 16),
     ]);
   }
 }