+++ tests/language_2/generic/metadata_in_function_body_test.dart
@@ -7,6 +7,8 @@
 // Check that annotations inside function bodies cannot use type arguments, but
 // can be raw.
 
+// @dart=2.11
+
 class C<T> {
   const C();
 }